cbitvector: use std::size_t for sizes and const
This replaces `int`, `uint32_t` and `uint64_t` types in the interface
and the implementation with `size_t`, which is the more appropriate type
and "can store the maximum size of a theoretically possible object of
any type" (https://en.cppreference.com/w/cpp/types/size_t).  Comparisons
between signed and unsigned variables are eliminated.
This also makes methods and parameters `const` if possible.