cbitvector: fix SetBits
`CBitVector::SetBits` was broken, it xored the vector with the input
with instead of setting the bits to new values.  This reimplements the
function using `CBitVector::SetBitNoMask` and new helper function
`GetArrayBit`.  However, probably this can be implemented in a more
performant way.

Thanks to Sebastian Stammler <stammler@cbs.tu-darmstadt.de> for
reporting and suggesting the fix.